backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langextract==1.0.2
pymilvus==2.3.4
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.4
python-dotenv==1.0.0
